chore: organize folders
