+++ src/routes/Projects.tsx
@@ -1,5 +1,5 @@
-import { Canvas } from "../components/Canvas"
 import { Suspense } from "react"
+import { Canvas } from "../components/Projects/Canvas"
 
 export const Projects = () => {
   return (
