+++ src/routes/Root.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import { Navbar } from "../components/Navbar";
+import { Navbar } from "../components/Main/Navbar";
 
 export default function Root() {
   return (
